+++ CHANGES.rst
@@ -16,7 +16,7 @@ Unreleased
 - Cutouts of ASDF data in FITS format now include embedded ASDF metadata in an "ASDF" extension within the FITS file for
   Python versions greater than or equal to 3.11. [#170]
 - Bugfix for getting the pixel location of given coordinates in ``ASDFCutout``. [#175]
-- New ``RomanSpectralCutout`` class and ``roman_spectral_cut`` function for creating spectral cutouts from Roman Space Telescope ASDF data. [#179]
+- New ``RomanSpectralSubset`` class and ``roman_spectral_subset`` function for creating spectral subsets from Roman Space Telescope ASDF data. [#179]
 
 Breaking Changes
 ^^^^^^^^^^^^^^^^